LayoutTests/imported/w3c/web-platform-tests/imagebitmap-renderingcontext/context-creation-with-alpha-expected.txt


FAIL Test that an ImageBitmapRenderingContext with alpha disabled makes the canvas opaque assert_array_equals: property 1, expected 127 but got 255
PASS Test that an ImageBitmapRenderingContext with alpha enabled preserves the alpha 
PASS Test that the 'alpha' context creation attribute is true by default 
